Remove init_prepare_finalize
