+++ vllm/model_executor/layers/quantization/modelopt.py
@@ -27,10 +27,10 @@ from vllm.model_executor.layers.quantization.utils.flashinfer_fp4_moe import (
     select_nvfp4_gemm_impl)
 from vllm.model_executor.layers.quantization.utils.flashinfer_utils import (
     FlashinferMoeBackend, apply_flashinfer_per_tensor_scale_fp8,
-    build_flashinfer_cutlass_moe_fp8_prepare_finalize,
-    get_flashinfer_moe_backend, register_moe_scaling_factors,
-    rotate_flashinfer_fp8_moe_weights, select_cutlass_fp8_gemm_impl,
-    swap_w13_to_w31)
+    build_flashinfer_fp8_cutlass_moe_prepare_finalize,
+    flashinfer_cutlass_moe_fp8, get_flashinfer_moe_backend,
+    register_moe_scaling_factors, rotate_flashinfer_fp8_moe_weights,
+    select_cutlass_fp8_gemm_impl, swap_w13_to_w31)
 from vllm.model_executor.layers.quantization.utils.marlin_utils_fp4 import (
     apply_fp4_marlin_linear, is_fp4_marlin_supported,
     prepare_fp4_layer_for_marlin, prepare_moe_fp4_layer_for_marlin)
@@ -300,9 +300,9 @@ class ModelOptFp8MoEMethod(FusedMoEMethodBase):
             self.flashinfer_moe_backend != FlashinferMoeBackend.CUTLASS:
             return super().maybe_make_prepare_finalize(moe)
 
-        prepare_finalize = build_flashinfer_cutlass_moe_fp8_prepare_finalize(
+        prepare_finalize = build_flashinfer_fp8_cutlass_moe_prepare_finalize(
             moe,
-            a1_gscale=self.layer.w13_input_scale,
+            layer=self.layer,
         )
         logger.debug_once("%s", prepare_finalize.__class__.__name__)
         return prepare_finalize
@@ -467,8 +467,6 @@ class ModelOptFp8MoEMethod(FusedMoEMethodBase):
             if self.flashinfer_moe_backend == FlashinferMoeBackend.TENSORRT_LLM:
                 rotate_flashinfer_fp8_moe_weights(layer.w13_weight,
                                                   layer.w2_weight)
-            elif self.flashinfer_moe_backend == FlashinferMoeBackend.CUTLASS:
-                self.init_prepare_finalize()
 
     def apply(
         self,
@@ -527,23 +525,34 @@ class ModelOptFp8MoEMethod(FusedMoEMethodBase):
         )
 
         if self.flashinfer_moe_backend == FlashinferMoeBackend.CUTLASS:
-            assert self.fused_experts is not None, (
-                "Expected self.fused_experts to be initialized")
             assert not renormalize
             assert activation == 'silu', (
                 f"Expected 'silu' activation but got {activation}")
-            return self.fused_experts(
-                x,
-                layer.w13_weight,
-                layer.w2_weight,
-                topk_weights,
-                topk_ids,
-                inplace=False,
-                activation=activation,
-                global_num_experts=global_num_experts,
-                expert_map=expert_map,
-                apply_router_weight_on_input=apply_router_weight_on_input,
-            )
+            if self.fused_experts is not None:
+                return self.fused_experts(
+                    x,
+                    layer.w13_weight,
+                    layer.w2_weight,
+                    topk_weights,
+                    topk_ids,
+                    inplace=False,
+                    activation=activation,
+                    global_num_experts=global_num_experts,
+                    expert_map=expert_map,
+                    apply_router_weight_on_input=apply_router_weight_on_input,
+                )
+            else:
+                return flashinfer_cutlass_moe_fp8(
+                    x,
+                    layer,
+                    topk_weights,
+                    topk_ids,
+                    inplace=False,
+                    activation=activation,
+                    global_num_experts=global_num_experts,
+                    expert_map=expert_map,
+                    apply_router_weight_on_input=apply_router_weight_on_input,
+                )
         from vllm.model_executor.layers.fused_moe.fused_moe import (
             fused_experts)
         return fused_experts(
